Not Just Gomoku V1
I made this when I was at a competition and had nothing to do. I pretty much abandoned it until a week ago where I decided I needed something to do while procrastinating. Do note that like any regular versioning, a change in the number is a big thing, and this being V1 is completely different from V2 onwards which I coded from the ground up.